feat: add rake command

+++ rakefile.rb
@@ -1,6 +1,13 @@
 require 'rake'
 require 'yaml'
 
+def ask(question, valid_options)
+  print question + " "
+  answer = STDIN.gets.chomp.downcase
+  return answer if valid_options.include?(answer)
+  abort("rake aborted!")
+end
+
 SOURCE = "."
 CONFIG = {
   'posts' => File.join(SOURCE, "_posts"),
@@ -21,11 +28,10 @@ task :post do
 
   puts "Creating new post: #{filename}"
 
-
   open(filename, 'w') do |post|
     post.puts "---"
     post.puts "layout: post"
-    post.puts "title: \"#{title.gsub(/-/,' ')}\""
+    post.puts 'title: "' + title.gsub(/-/,' ') + '"'
     # post.puts "subtitle: \"\""
     post.puts "date: #{Time.now.strftime('%Y-%m-%d')}"
     # post.puts "cover: "
@@ -33,8 +39,6 @@ task :post do
     post.puts "tags: "
     post.puts "---"
   end
-
-  
 end # task :post
 
 ## 第二个命令
@@ -50,14 +54,54 @@ task :draft do
 
   puts "Creating new post: #{filename}"
 
+  open(filename, 'w') do |post|
+    post.puts "---"
+    post.puts "layout: post"
+    post.puts 'title: "' + title.gsub(/-/,' ') + '"'
+    # post.puts "subtitle: \"\""
+    post.puts "date: #{Time.now.strftime('%Y-%m-%d')}"
+    # post.puts "cover: "
+    post.puts "category: "
+    post.puts "tags: "
+    post.puts "---"
+  end
+end
+
+#-- Tasks for starting the server and creating new posts
+
+desc "Serve the site locally with livereload"
+task :serve do
+  puts "Starting Jekyll server..."
+  system("bundle exec jekyll serve --livereload --port 4000")
+end
+
+task :start => :serve
+
+desc "Create a new post with an interactive prompt"
+task :new_post do
+  print "Enter post title: "
+  title = STDIN.gets.chomp
+  
+  if title.strip.empty?
+    abort("rake aborted: title cannot be empty.")
+  end
+
+  slug = title.downcase.strip.gsub(' ', '-').gsub(/[^\w-]/, '')
+  filename = File.join(CONFIG['posts'], "#{Time.now.strftime('%Y-%m-%d')}-#{slug}.#{CONFIG['post_ext']}")
+
+  if File.exist?(filename)
+    print "#{filename} already exists. Overwrite? [y/n]: "
+    overwrite = STDIN.gets.chomp.downcase
+    abort("rake aborted!") unless overwrite == 'y'
+  end
+
+  puts "Creating new post: #{filename}"
 
   open(filename, 'w') do |post|
     post.puts "---"
     post.puts "layout: post"
-    post.puts "title: \"#{title.gsub(/-/,' ')}\""
-    # post.puts "subtitle: \"\""
-    post.puts "date: #{Time.now.strftime('%Y-%m-%d')}"
-    # post.puts "cover: "
+    post.puts 'title: "' + title + '"'
+    post.puts "date: #{Time.now.strftime('%Y-%m-%d %H:%M:%S %z')}"
     post.puts "category: "
     post.puts "tags: "
     post.puts "---"
